+++ test/Config/Option/Tracking/DisableReferrerTrackingConfigOptionTest.php
@@ -2,29 +2,29 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Installer\Config\Option\UrlShortener;
+namespace ShlinkioTest\Shlink\Installer\Config\Option\Tracking;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Shlinkio\Shlink\Config\Collection\PathCollection;
-use Shlinkio\Shlink\Installer\Config\Option\UrlShortener\OrphanVisitsTrackingConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Tracking\DisableReferrerTrackingConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class OrphanVisitsTrackingConfigOptionTest extends TestCase
+class DisableReferrerTrackingConfigOptionTest extends TestCase
 {
     use ProphecyTrait;
 
-    private OrphanVisitsTrackingConfigOption $configOption;
+    private DisableReferrerTrackingConfigOption $configOption;
 
     public function setUp(): void
     {
-        $this->configOption = new OrphanVisitsTrackingConfigOption();
+        $this->configOption = new DisableReferrerTrackingConfigOption();
     }
 
     /** @test */
     public function returnsExpectedConfig(): void
     {
-        self::assertEquals(['url_shortener', 'track_orphan_visits'], $this->configOption->getConfigPath());
+        self::assertEquals(['tracking', 'disable_referrer_tracking'], $this->configOption->getConfigPath());
     }
 
     /** @test */
@@ -32,10 +32,7 @@ class OrphanVisitsTrackingConfigOptionTest extends TestCase
     {
         $expectedAnswer = true;
         $io = $this->prophesize(StyleInterface::class);
-        $confirm = $io->confirm(
-            'Do you want track orphan visits? (visits to the base URL, invalid short URLs or other "not found" URLs)',
-            true,
-        )->willReturn(
+        $confirm = $io->confirm('Do you want to disable tracking of visitors\' "Referrers"?', false)->willReturn(
             $expectedAnswer,
         );